fix: address second Copilot review — TS narrowing, dead ref, encodeURIComponent, test file split
- Assign NEXT_PUBLIC_ESRI_API_KEY to a local const so TypeScript narrows
  the type from string|undefined to string before calling esriTileUrl
- Remove internalMapRef (useRef) — it was assigned in onReady but never
  read after the rAF zoom logic was removed
- Wrap API key in encodeURIComponent to handle keys with +, /, = etc.
- Move esriTileUrl tests from Map.test.tsx into a dedicated esriTileUrl.test.ts
  so test intent and file naming are aligned; add a URL-encoding test case
